ai-detection-dashboard2/ai-detection-dashboard/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Werkzeug==2.3.7
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
Pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2
reportlab==4.0.4
python-dotenv==1.0.0
WTForms==3.0.1
gunicorn==21.2.0
redis==4.6.0
celery==5.3.1
cryptography==41.0.3
hashlib-additional==1.0.1